Rename web models package for consistency
