+++ app/build.gradle
@@ -37,7 +37,7 @@ dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     compile "com.android.support:appcompat-v7:$project.supportVersion"
     compile "com.android.support:cardview-v7:$project.supportVersion"
-    compile "com.android.support.constraint:constraint-layout:1.0.2"
+    compile 'com.android.support.constraint:constraint-layout:1.0.2'
     compile "com.android.support:design:$project.supportVersion"
     compile "com.android.support:recyclerview-v7:$project.supportVersion"
     compile "com.android.support:support-v4:$project.supportVersion"
@@ -46,8 +46,8 @@ dependencies {
     compile 'com.github.armcha:AutoLinkTextView:0.3.0'
 
     // networking
-    compile "com.google.code.gson:gson:2.8.1"
-    compile "com.squareup.okhttp3:okhttp:3.7.0"
+    compile 'com.google.code.gson:gson:2.8.1'
+    compile 'com.squareup.okhttp3:okhttp:3.7.0'
     compile 'com.squareup.picasso:picasso:2.5.2'
     compile 'com.jakewharton.picasso:picasso2-okhttp3-downloader:1.1.0'
 
@@ -58,7 +58,7 @@ dependencies {
     compile 'joda-time:joda-time:2.9.9'
 
     // logger
-    compile "com.jakewharton.timber:timber:4.5.1"
+    compile 'com.jakewharton.timber:timber:4.5.1'
 
     // dependency injection
     compile project.daggerAndroidSupport, {
@@ -68,10 +68,10 @@ dependencies {
     annotationProcessor project.daggerAnnotationProcessor
 
     // test libraries
-    testCompile "junit:junit:4.12"
-    testCompile "org.mockito:mockito-core:2.8.9"
-    testCompile "org.robolectric:robolectric:3.4.2"
-    testCompile "org.robolectric:shadows-support-v4:3.0"
+    testCompile 'junit:junit:4.12'
+    testCompile 'org.mockito:mockito-core:2.8.9'
+    testCompile 'org.robolectric:robolectric:3.4.2'
+    testCompile 'org.robolectric:shadows-support-v4:3.0'
 
     // instrumented test libraries
     androidTestCompile "com.android.support:support-annotations:$project.supportVersion"
