+++ app/src/main/java/com/github/fo2rist/mclaren/repository/McLarenFeedRepository.java
@@ -6,9 +6,9 @@ import com.github.fo2rist.mclaren.models.FeedItem;
 import com.github.fo2rist.mclaren.web.FeedHistoryPredictor;
 import com.github.fo2rist.mclaren.web.FeedWebsevice;
 import com.github.fo2rist.mclaren.web.WebCallback;
-import com.github.fo2rist.mclaren.web.model.McLarenFeed;
-import com.github.fo2rist.mclaren.web.model.McLarenFeedConverter;
-import com.github.fo2rist.mclaren.web.model.ResponseParser;
+import com.github.fo2rist.mclaren.web.models.McLarenFeed;
+import com.github.fo2rist.mclaren.web.models.McLarenFeedConverter;
+import com.github.fo2rist.mclaren.web.models.ResponseParser;
 import java.io.IOException;
 import java.net.URL;
 import java.util.ArrayList;
