+++ app/src/main/java/com/github/fo2rist/mclaren/web/models/McLarenMediaItem.java
@@ -1,4 +1,4 @@
-package com.github.fo2rist.mclaren.web.model;
+package com.github.fo2rist.mclaren.web.models;
 
 import android.support.annotation.Nullable;
